HTMGT: OCC Initialization
Contains the changes for p9 OCC init. Mostly
a port from p8. New cmd/rsp buffers and data format
for packets. This commit also makes HTMGT fapi2
compatible.

RTC:133945
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/31010
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/usr/ipmi/ipmisensor.C
@@ -1156,7 +1156,7 @@ namespace SENSOR
 
     // interface to retrieve the APSS channel sensor numbers.
     errlHndl_t getAPSSChannelSensorNumbers(
-            const uint16_t (* &o_sensor_numbers)[16])
+            const uint32_t (* &o_sensor_numbers)[16])
     {
 
         TARGETING::TargetHandle_t l_sys;
